Keyword highlighting improvements
This uses match rules instead of keyword rules to allow Vim to highlight
reserved keywords even when there's a numeric literal immediately
preceding with no whitespace (which is valid).

This also avoids an issue where the RHS of a sequence would get
highlighted as if it were a float literal.